Move types and helpers into topical files
